move checkSession function to its own util file
